Move utility classes from styx-api into styx-common.
