--- components/api/src/main/java/com/hotels/styx/api/common/Strings.java
@@ -1,46 +1,0 @@
-/*
-  Copyright (C) 2013-2018 Expedia Inc.
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
- */
-package com.hotels.styx.api.common;
-
-/**
- * Utility functions for working with String objects.
- */
-public final class Strings {
-    private Strings() {
-    }
-
-    /**
-     * Return true if the string {@code suffix} is null or if name ends with {@code suffix}.
-     *
-     * @param suffix a suffix
-     * @param name   a name
-     * @return true or false
-     */
-    public static boolean hasSuffix(String suffix, String name) {
-        return suffix == null || name.endsWith(suffix);
-    }
-
-    /**
-     * Wraps a string value with double qoutes.
-     *
-     * @param value
-     * @return the quoted value
-     */
-    public static String quote(String value) {
-        return "\"" + value + "\"";
-    }
-
-}
