+++ components/common/src/test/java/com/hotels/styx/common/MorePreconditionsTest.java
@@ -13,12 +13,12 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.common;
+package com.hotels.styx.common;
 
 import org.testng.annotations.Test;
 
-import static com.hotels.styx.api.common.MorePreconditions.checkArgument;
-import static com.hotels.styx.api.common.MorePreconditions.inRange;
+import static com.hotels.styx.common.MorePreconditions.checkArgument;
+import static com.hotels.styx.common.MorePreconditions.inRange;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.is;
